@@ -5,35 +5,24 @@
   "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --config webpack.config.prod.js",
-    "serve": "webpack-dev-server"
+    "build": "parcel build app/index.html",
+    "run": "parcel app/index.html"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-cli": "^6.10.1",
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-lodash": "^3.2.0",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^2.13.1",
-    "eslint-plugin-react": "^5.2.2",
-    "file-loader": "^0.9.0",
-    "json-loader": "^0.5.4",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-polyfill": "^6.26.0",
+    "bootstrap": "^3.3.7",
+    "eslint": "^5.7.0",
+    "jquery": "^3.3.1",
+    "lodash": "^4.17.11",
+    "parcel-bundler": "^1.10.3",
+    "prop-types": "^15.6.2",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2"
   },
-  "dependencies": {
-    "bootstrap": "^3.3.6",
-    "jquery": "^3.0.0",
-    "lodash": "^4.13.1",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0"
-  }
+  "dependencies": {}
 }
